[monitoring/rid_qualifier] Rename rid_qualifier package to uss_qualifier
