+++ monitoring/uss_qualifier/Dockerfile
