+++ monitoring/uss_qualifier/mock/Dockerfile
