+++ monitoring/uss_qualifier/rid_qualifier_entry.py
@@ -6,8 +6,8 @@ import sys
 import argparse
 from urllib.parse import urlparse
 from monitoring.monitorlib.typing import ImplicitDict
-from monitoring.rid_qualifier.utils import RIDQualifierTestConfiguration
-import monitoring.rid_qualifier.test_executor as test_executor
+from monitoring.uss_qualifier.utils import RIDQualifierTestConfiguration
+import monitoring.uss_qualifier.test_executor as test_executor
 
 def is_url(url_string):
     try:
