+++ monitoring/uss_qualifier/run_flight_state_locally.sh
