+++ monitoring/uss_qualifier/webapp/resources.py
@@ -1,6 +1,6 @@
 import rq
 from redis import Redis
-from monitoring.rid_qualifier.webapp import webapp
+from monitoring.uss_qualifier.webapp import webapp
 from . import config
