+++ monitoring/uss_qualifier/webapp/tasks.py
@@ -3,10 +3,10 @@ import redis
 import rq
 from . import resources
 from monitoring.monitorlib.typing import ImplicitDict
-from monitoring.rid_qualifier import test_executor
-from monitoring.rid_qualifier.utils import RIDQualifierTestConfiguration
-from monitoring.rid_qualifier import create_flight_record_from_kml
-from monitoring.rid_qualifier.test_data import test_report
+from monitoring.uss_qualifier import test_executor
+from monitoring.uss_qualifier.utils import RIDQualifierTestConfiguration
+from monitoring.uss_qualifier import create_flight_record_from_kml
+from monitoring.uss_qualifier.test_data import test_report
 
 
 def get_rq_job(job_id):
